requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
Django==3.0.3
django-bootstrap4==22.1
django-crispy-forms==1.14.0
djangorestframework==3.13.1
pytz==2022.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
